Replace TaskSequenceHandler interface with concrete implementation
