+++ ground/src/main/java/com/google/android/ground/ui/datacollection/TaskSequenceHandler.kt
@@ -18,7 +18,31 @@ package com.google.android.ground.ui.datacollection
 import com.google.android.ground.model.submission.TaskData
 import com.google.android.ground.model.task.Task
 
-interface TaskSequenceHandler {
+/** Handles the operations related to task sequences. */
+class TaskSequenceHandler(
+  private val tasks: List<Task>,
+  private val shouldIncludeTask:
+    (task: Task, taskValueOverride: Pair<String, TaskData?>?) -> Boolean,
+) {
+
+  init {
+    if (tasks.isEmpty()) {
+      error("Can't generate sequence for empty task list")
+    }
+  }
+
+  private fun checkInvalidTaskId(taskId: String) {
+    if (taskId.isBlank()) {
+      error("Task ID can't be blank")
+    }
+  }
+
+  private fun checkInvalidIndex(taskId: String, index: Int) {
+    checkInvalidTaskId(taskId)
+    if (index < 0) {
+      error("Task $taskId not found in task list")
+    }
+  }
 
   /**
    * Retrieves the task sequence based on the provided inputs and conditions.
@@ -30,7 +54,8 @@ interface TaskSequenceHandler {
    *   element is the [TaskData] to override the default task data. If null, no override is applied.
    * @return A [Sequence] of [Task] objects representing the ordered tasks.
    */
-  fun getTaskSequence(taskValueOverride: Pair<String, TaskData?>? = null): Sequence<Task>
+  fun getTaskSequence(taskValueOverride: Pair<String, TaskData?>? = null): Sequence<Task> =
+    tasks.filter { task -> shouldIncludeTask(task, taskValueOverride) }.asSequence()
 
   /**
    * Checks if the specified task is the first task in the displayed sequence.
@@ -38,7 +63,10 @@ interface TaskSequenceHandler {
    * @param taskId The ID of the task to check.
    * @return `true` if the task is the first in the sequence, `false` otherwise.
    */
-  fun isFirstPosition(taskId: String): Boolean
+  fun isFirstPosition(taskId: String): Boolean {
+    checkInvalidTaskId(taskId)
+    return taskId == getTaskSequence().first().id
+  }
 
   /**
    * Checks if the specified task is the last task in the displayed sequence.
@@ -46,7 +74,10 @@ interface TaskSequenceHandler {
    * @param taskId The ID of the task to check.
    * @return `true` if the task is the last in the sequence, `false` otherwise.
    */
-  fun isLastPosition(taskId: String): Boolean
+  fun isLastPosition(taskId: String): Boolean {
+    checkInvalidTaskId(taskId)
+    return taskId == getTaskSequence().last().id
+  }
 
   /**
    * Checks if the specified task with the given data is the last task in the sequence.
@@ -57,7 +88,10 @@ interface TaskSequenceHandler {
    * @param value The [TaskData] associated with the task.
    * @return `true` if the task with the given data is the last in the sequence, `false` otherwise.
    */
-  fun isLastPosition(taskId: String, value: TaskData?): Boolean
+  fun isLastPosition(taskId: String, value: TaskData?): Boolean {
+    checkInvalidTaskId(taskId)
+    return taskId == getTaskSequence(taskValueOverride = taskId to value).last().id
+  }
 
   /**
    * Retrieves the ID of the task that precedes the specified task in the sequence.
@@ -65,7 +99,10 @@ interface TaskSequenceHandler {
    * @param taskId The ID of the task for which to find the previous task.
    * @return The ID of the previous task in the sequence.
    */
-  fun getPreviousTask(taskId: String): String
+  fun getPreviousTask(taskId: String): String {
+    checkInvalidTaskId(taskId)
+    return getTaskSequence().elementAt(getRelativePosition(taskId) - 1).id
+  }
 
   /**
    * Retrieves the ID of the task that follows the specified task in the sequence.
@@ -73,7 +110,10 @@ interface TaskSequenceHandler {
    * @param taskId The ID of the task for which to find the next task.
    * @return The ID of the next task in the sequence.
    */
-  fun getNextTask(taskId: String): String
+  fun getNextTask(taskId: String): String {
+    checkInvalidTaskId(taskId)
+    return getTaskSequence().elementAt(getRelativePosition(taskId) + 1).id
+  }
 
   /**
    * Retrieves the absolute index of the specified task in the original, unfiltered task sequence.
@@ -84,7 +124,11 @@ interface TaskSequenceHandler {
    * @param taskId The ID of the task for which to find the absolute position.
    * @throws error if the task is not found in the original task sequence.
    */
-  fun getAbsolutePosition(taskId: String): Int
+  fun getAbsolutePosition(taskId: String): Int {
+    val index = tasks.indexOfFirst { it.id == taskId }
+    checkInvalidIndex(taskId, index)
+    return index
+  }
 
   /**
    * Retrieves the relative index in the computed task sequence.
@@ -94,12 +138,23 @@ interface TaskSequenceHandler {
    * @param taskId The ID of the task for which to find the relative position.
    * @throws error if the task is not found in the computed task sequence.
    */
-  fun getRelativePosition(taskId: String): Int
+  fun getRelativePosition(taskId: String): Int {
+    val index = getTaskSequence().indexOfFirst { it.id == taskId }
+    checkInvalidIndex(taskId, index)
+    return index
+  }
 
   /**
    * Retrieves the position information for the specified task.
    *
    * @param taskId THE ID of the task for which to find the position information.
    */
-  fun getTaskPosition(taskId: String): TaskPosition
+  fun getTaskPosition(taskId: String): TaskPosition {
+    checkInvalidTaskId(taskId)
+    return TaskPosition(
+      absoluteIndex = getAbsolutePosition(taskId),
+      relativeIndex = getRelativePosition(taskId),
+      sequenceSize = getTaskSequence().toList().size,
+    )
+  }
 }
